requirements.txt
absl-py==1.0.0
AMP==1.1.4
anykeystore==0.2
apex==0.9.10.dev0
asgiref==3.5.0
cached-property==1.5.2
cachetools==4.2.4
caldav==0.1.4
certifi==2021.10.8
charset-normalizer==2.0.9
coverage==6.3.2
cryptacular==1.6.2
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.12
fonttools==4.28.5
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.42.0
h5py==3.6.0
hupper==1.10.3
idna==3.3
imageio==2.16.0
importlib-metadata==4.8.2
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
networkx==2.6.3
nose==1.3.7
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634106693478/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.4.60
packaging==21.3
pandas==1.1.5
PasteDeploy==2.1.1
pbkdf2==1.3
Pillow==8.4.0
plaster==1.0
plaster-pastedeploy==0.7
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pyramid==2.0
pyramid-mailer==0.15.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyWavelets==1.2.0
PyYAML==5.4.1
repoze.sendmail==4.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.1
scipy==1.7.3
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
str2bool==1.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.8.2
tqdm==4.62.3
transaction==3.0.1
translationstring==1.4
ttach==0.0.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
velruse==1.1.1
venusian==3.0.0
vobject==0.9.6.1
WebOb==1.8.7
Werkzeug==2.0.2
WTForms==3.0.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
zipp==3.6.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
